Avoid implicit float to int conversions

+++ gtk2_ardour/ardour_ui.cc
@@ -1416,7 +1416,7 @@ ARDOUR_UI::update_disk_space()
 		/* Available space is unknown */
 		format_disk_space_label (-1);
 	} else if (opt_samples.value_or (0) == max_samplecnt) {
-		format_disk_space_label (max_samplecnt);
+		format_disk_space_label (static_cast<float>(max_samplecnt));
 	} else {
 		rec_enabled_streams = 0;
 		_session->foreach_route (this, &ARDOUR_UI::count_recenabled_streams, false);
